API query still unresolved
However, I have re-organised the file system for clarity and added
styling to the SearchBar component. Our current theory is that we may
need to switch API provider.
