+++ fetch/src/app/components/Footer/footer.tsx
