+++ fetch/src/app/components/SearchBar/SearchBar.module.css
@@ -1,0 +1,58 @@
+.searchBar {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 20px;
+    background-color: #f1f8f5; /* light sage green background */
+    border-radius: 8px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    max-width: 400px;
+    margin: 0 auto;
+}
+
+.input {
+    width: 100%;
+    padding: 12px;
+    margin-bottom: 10px;
+    border: 2px solid #a8d8c8; 
+    border-radius: 4px;
+    font-size: 16px;
+    color: #3c5a4e; 
+    background-color: #ffffff;
+    outline: none;
+    transition: border-color 0.3s ease-in-out;
+}
+
+.input:focus {
+    border-color: #5b9f8c; 
+}
+
+.button {
+    padding: 12px 20px;
+    font-size: 16px;
+    color: darkgreen;
+    font-weight: bold;
+    background-color: rgb(192, 228, 192); 
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s ease;
+    width: 100%;
+}
+
+.button:disabled {
+    background-color: #b0d2c5; 
+    cursor: not-allowed;
+}
+
+.button:hover:not(:disabled) {
+    background-color: #5b9f8c; 
+}
+
+.error {
+    color: #e74c3c; 
+    margin-top: 10px;
+    font-size: 14px;
+    font-weight: bold;
+}
